seperated class libray and console app
