+++ src/MathGame.ConsoleApplication/MathGame.ConsoleApplication.csproj
@@ -7,4 +7,8 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\MathGame\MathGame.csproj" />
+  </ItemGroup>
+
 </Project>
