+++ src/MathGame.sln
@@ -5,7 +5,7 @@ VisualStudioVersion = 17.8.34330.188
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MathGame", "MathGame\MathGame.csproj", "{1958C166-27AA-4FE6-A735-30227C699F89}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MathGame.Console", "MathGame.Console\MathGame.Console.csproj", "{2F2B5043-3A93-4091-BF9D-CC78F8429C30}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MathGame.ConsoleApplication", "MathGame.ConsoleApplication\MathGame.ConsoleApplication.csproj", "{4914E134-125A-403F-BAF0-5F9CEB94F412}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -17,10 +17,10 @@ Global
 		{1958C166-27AA-4FE6-A735-30227C699F89}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1958C166-27AA-4FE6-A735-30227C699F89}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1958C166-27AA-4FE6-A735-30227C699F89}.Release|Any CPU.Build.0 = Release|Any CPU
-		{2F2B5043-3A93-4091-BF9D-CC78F8429C30}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2F2B5043-3A93-4091-BF9D-CC78F8429C30}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2F2B5043-3A93-4091-BF9D-CC78F8429C30}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2F2B5043-3A93-4091-BF9D-CC78F8429C30}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4914E134-125A-403F-BAF0-5F9CEB94F412}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4914E134-125A-403F-BAF0-5F9CEB94F412}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4914E134-125A-403F-BAF0-5F9CEB94F412}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4914E134-125A-403F-BAF0-5F9CEB94F412}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
