--- src/MathGame/Program.cs
@@ -1,15 +1,0 @@
-﻿// -------------------------------------------------------------------------------------------------
-// MathGame.Program
-// -------------------------------------------------------------------------------------------------
-// Add some description
-// -------------------------------------------------------------------------------------------------
-using MathGame;
-
-var menu = new Menu();
-var date = DateTime.UtcNow;
-
-var games = new List<string>();
-
-string name = Helpers.GetName();
-
-menu.ShowMenu(name, date);
